Move json equality logic outside of test_arrow
